Code cleanup; move upgrade test to new folder
