+++ tests/integration/helm/install_test.go
@@ -16,51 +16,16 @@
 package helm
 
 import (
-	"context"
 	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
-	v1 "k8s.io/api/core/v1"
-	kerrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-
-	"istio.io/istio/pkg/test/env"
 	"istio.io/istio/pkg/test/framework"
 	"istio.io/istio/pkg/test/framework/components/environment/kube"
 	"istio.io/istio/pkg/test/framework/image"
-	"istio.io/istio/pkg/test/framework/resource"
 	"istio.io/istio/pkg/test/helm"
-	kubetest "istio.io/istio/pkg/test/kube"
-	"istio.io/istio/pkg/test/scopes"
-	"istio.io/istio/pkg/test/util/retry"
-	"istio.io/pkg/log"
-)
-
-const (
-	IstioNamespace      = "istio-system"
-	ReleasePrefix       = "istio-"
-	BaseChart           = "base"
-	DiscoveryChart      = "istio-discovery"
-	IngressGatewayChart = "istio-ingress"
-	EgressGatewayChart  = "istio-egress"
-	BaseReleaseName     = ReleasePrefix + BaseChart
-	IstiodReleaseName   = "istiod"
-	IngressReleaseName  = IngressGatewayChart
-	EgressReleaseName   = EgressGatewayChart
-	ControlChartsDir    = "istio-control"
-	GatewayChartsDir    = "gateways"
-	retryDelay          = 2 * time.Second
-	retryTimeOut        = 5 * time.Minute
-	helmTimeout         = 2 * time.Minute
-)
-
-var (
-	// ChartPath is path of local Helm charts used for testing.
-	ChartPath = filepath.Join(env.IstioSrc, "manifests/charts")
 )
 
 // TestDefaultInstall tests Istio installation using Helm with default options
@@ -89,9 +54,9 @@ global:
 			if err := ioutil.WriteFile(overrideValuesFile, []byte(overrideValues), os.ModePerm); err != nil {
 				t.Fatalf("failed to write iop cr file: %v", err)
 			}
-			installGatewaysCharts(t, cs, h, overrideValuesFile)
+			InstallGatewaysCharts(t, cs, h, overrideValuesFile)
 
-			verifyInstallation(ctx, cs)
+			VerifyInstallation(ctx, cs)
 
 			t.Cleanup(func() {
 				deleteGatewayCharts(t, h)
@@ -127,88 +92,12 @@ global:
 			if err := ioutil.WriteFile(overrideValuesFile, []byte(overrideValues), os.ModePerm); err != nil {
 				t.Fatalf("failed to write iop cr file: %v", err)
 			}
-			installGatewaysCharts(t, cs, h, overrideValuesFile)
+			InstallGatewaysCharts(t, cs, h, overrideValuesFile)
 
-			verifyInstallation(ctx, cs)
+			VerifyInstallation(ctx, cs)
 
 			t.Cleanup(func() {
 				deleteGatewayCharts(t, h)
 			})
 		})
-}
-
-// installGatewaysCharts install Istio using Helm charts with the provided
-// override values file and fails the tests on any failures.
-func installGatewaysCharts(t *testing.T, cs resource.Cluster,
-	h *helm.Helm, overrideValuesFile string) {
-	createIstioSystemNamespace(t, cs)
-
-	// Install ingress gateway chart
-	err := h.InstallChart(IngressReleaseName, filepath.Join(GatewayChartsDir, IngressGatewayChart),
-		IstioNamespace, overrideValuesFile, helmTimeout)
-	if err != nil {
-		t.Errorf("failed to install istio %s chart", IngressGatewayChart)
-	}
-
-	// Install egress gateway chart
-	err = h.InstallChart(EgressReleaseName, filepath.Join(GatewayChartsDir, EgressGatewayChart),
-		IstioNamespace, overrideValuesFile, helmTimeout)
-	if err != nil {
-		t.Errorf("failed to install istio %s chart", EgressGatewayChart)
-	}
-}
-
-func createIstioSystemNamespace(t *testing.T, cs resource.Cluster) {
-	if _, err := cs.CoreV1().Namespaces().Create(context.TODO(), &v1.Namespace{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: IstioNamespace,
-		},
-	}, metav1.CreateOptions{}); err != nil {
-		if kerrors.IsAlreadyExists(err) {
-			log.Debugf("%v namespace already exist", IstioNamespace)
-		} else {
-			t.Errorf("failed to create %v namespace: %v", IstioNamespace, err)
-		}
-	}
-}
-
-// deleteGatewayCharts deletes installed Istio Helm charts and resources
-func deleteGatewayCharts(t *testing.T, h *helm.Helm) {
-	scopes.Framework.Infof("cleaning up resources")
-	if err := h.DeleteChart(EgressReleaseName, IstioNamespace); err != nil {
-		t.Errorf("failed to delete %s release", EgressReleaseName)
-	}
-	if err := h.DeleteChart(IngressReleaseName, IstioNamespace); err != nil {
-		t.Errorf("failed to delete %s release", IngressReleaseName)
-	}
-}
-
-// verifyInstallation verify that the Helm installation is successful
-func verifyInstallation(ctx framework.TestContext, cs resource.Cluster) {
-	scopes.Framework.Infof("=== verifying istio installation === ")
-
-	retry.UntilSuccessOrFail(ctx, func() error {
-		if _, err := kubetest.CheckPodsAreReady(kubetest.NewPodFetch(cs, IstioNamespace, "app=istiod")); err != nil {
-			return fmt.Errorf("istiod pod is not ready: %v", err)
-		}
-		if _, err := kubetest.CheckPodsAreReady(kubetest.NewPodFetch(cs, IstioNamespace, "app=istio-ingressgateway")); err != nil {
-			return fmt.Errorf("istio ingress gateway pod is not ready: %v", err)
-		}
-		if _, err := kubetest.CheckPodsAreReady(kubetest.NewPodFetch(cs, IstioNamespace, "app=istio-egressgateway")); err != nil {
-			return fmt.Errorf("istio egress gateway pod is not ready: %v", err)
-		}
-		return nil
-	}, retry.Timeout(retryTimeOut), retry.Delay(retryDelay))
-	scopes.Framework.Infof("=== succeeded ===")
-}
-
-func getValuesOverrides(ctx framework.TestContext, valuesStr, hub, tag string) string {
-	workDir := ctx.CreateTmpDirectoryOrFail("helm")
-	overrideValues := fmt.Sprintf(valuesStr, hub, tag)
-	overrideValuesFile := filepath.Join(workDir, "values.yaml")
-	if err := ioutil.WriteFile(overrideValuesFile, []byte(overrideValues), os.ModePerm); err != nil {
-		ctx.Fatalf("failed to write iop cr file: %v", err)
-	}
-
-	return overrideValuesFile
 }
